# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Flowstile API
  version: 0.1.0
  description: >-
    Human-task inbox and form layer for Temporal.io workflows.


    Authentication uses JWT tokens delivered as HttpOnly cookies (`flowstile_token`) or as Bearer tokens in the
    Authorization header.
  license:
    name: Apache-2.0
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: flowstile_token
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  pattern: ^(?!\.)(?!.*\.\.)([A-Za-z0-9_'+\-\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
                password:
                  type: string
                  minLength: 1
              required:
                - email
                - password
      responses:
        "200":
          description: Default Response
...
  /auth/api-keys:
    post:
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                permissions:
                  minItems: 1
                  type: array
                  items:
                    type: string
                    enum:
                      - tasks:read
                      - tasks:write
                      - tasks:manage
                      - forms:write
                      - processes:write
                      - users:manage
                expiresAt: {}
              required:
                - name
                - permissions
      responses:
        "200":
          description: Default Response
    get:
      tags:
        - Auth
      responses:
        "200":
          description: Default Response
  /auth/api-keys/{id}:
    delete:
      tags:
        - Auth
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
...
  /tasks/{id}/attachments:
    post:
      tags:
        - Attachments
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
  /tasks/{id}/attachments/{attachmentId}/content:
    get:
      tags:
        - Attachments
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: attachmentId
          required: true
      responses:
        "200":
          description: Default Response
...
security:
  - cookieAuth: []
  - bearerAuth: []
...